// src/orders/interfaces/order.interface.ts
import { OrderStatus } from 'generated/prisma/enums';

export interface OrderWithProducts {
  orderItems: {
    name: string;
    productId: number;
    quantity: number;
    price: number;
  }[];
  id: string;
  totalAmount: number;
  totalItems: number;
  status: OrderStatus;
  paid: boolean;
  paidAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
}